Prepares the Perakim components for splitting Torah off from them
